feat:补充wait4的逻辑，添加对进程组的判断

+++ kernel/src/process/syscall.rs
@@ -203,7 +203,7 @@ impl Syscall {
     }
 
     pub fn wait4(
-        pid: i64,
+        pid: i32,
         wstatus: *mut i32,
         options: i32,
         rusage: *mut c_void,
